Route overwriting fix
Small change that allows the route to be overwritten. Current behavior:
F.route('/',view_index);   
F.route('/',view_index2);

Even though we're specifying route for the 2nd time with view_index2, view_index 1 will always be used. Routes keep getting pushed into an array and right now it's 1st come 1st serve. So this fix checks if route with that destination is already in place and overwrites it.